Move FlashListProps to a separate file
